:truck: Move get_wallet_type check to own util module
