+++ app/services/definitions/credential_definitions.py
@@ -17,8 +17,8 @@ from app.services.definitions.credential_definition_publisher import (
 from app.services.trust_registry.util.issuer import assert_valid_issuer
 from app.util.assert_public_did import assert_public_did
 from app.util.definitions import credential_definition_from_acapy
-from app.util.tenants import get_wallet_type
 from app.util.transaction_acked import wait_for_transaction_ack
+from app.util.wallet_type_checks import get_wallet_type
 from shared import CRED_DEF_ACK_TIMEOUT, REGISTRY_SIZE
 from shared.log_config import get_logger
