+++ app/services/revocation_registry.py
@@ -24,7 +24,7 @@ from app.exceptions import (
 from app.models.issuer import ClearPendingRevocationsResult, RevokedResponse
 from app.util.credentials import strip_protocol_prefix
 from app.util.retry_method import coroutine_with_retry
-from app.util.tenants import get_wallet_type
+from app.util.wallet_type_checks import get_wallet_type
 from shared.log_config import get_logger
 
 logger = get_logger(__name__)
